refactor(policy): split policy.go into normalize.go and clean.go
Move OTel normalization helpers to normalize.go, cleaning helpers to
clean.go, and split the test file to match.
